refactor Device class for improved readability. Clean up includes and adjust formatting in various files. Remove unused discover files. Add .clang-format file listing rules for code formatting.

+++ server/SensorReading.hpp
@@ -1,8 +1,8 @@
 #pragma once
 struct SensorReading {
-    int value;
-    float min_value;
-    float max_value;
-    double sum;
-    unsigned int times;
+  float value;
+  float min_value;
+  float max_value;
+  double sum;
+  unsigned int times;
 };
